Extract `typing` module into `typarena` crate
Move the `typing` module (`ColumnId`, `TypePool`, `TypeTable`
and their backing maps) out of `fynix` into a standalone `no_std`
crate, `typarena`, so it can later live in its own repo. `fynix`
now depends on `typarena` and references it via `typarena::`
instead of `crate::typing::`.
