+++ crates/fynix/src/element/type_meta.rs
@@ -1,8 +1,8 @@
 use alloc::vec::Vec;
 
 use crate::element::{Element, ElementId};
-use crate::typing::ColumnId;
-use crate::typing::type_pool::TypePool;
+use typarena::ColumnId;
+use typarena::type_pool::TypePool;
 
 /// Per-type dispatch table registry, column-indexed.
 pub struct ElementTypeMetas {
