+++ crates/fynix/src/interaction.rs
@@ -1,6 +1,6 @@
 use crate::element::ElementId;
 use crate::events::Events;
-use crate::typing::type_table::TypeTable;
+use typarena::type_table::TypeTable;
 
 /// User-written interaction handler: reacts to interaction `I`,
 /// emitting messages into [`Events`].
