+++ crates/fynix/src/reactive.rs
@@ -8,7 +8,7 @@ use crate::element::layout::ElementNodes;
 use crate::element::{Element, ElementBuild, ElementId};
 use crate::init::Init;
 use crate::style::StyleId;
-use crate::typing::type_pool::{PoolKey, TypePool};
+use typarena::type_pool::{PoolKey, TypePool};
 
 pub struct Reactives {
     reactives: TypePool,
