+++ crates/typarena/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "typarena"
+description = "Type-keyed arena storage with stable per-type columns."
+keywords = ["arena", "type-map", "storage", "no_std"]
+categories = ["data-structures", "no-std"]
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+hashbrown.workspace = true
+sparse_map.workspace = true
+indexmap.workspace = true
