+++ crates/typarena/src/lib.rs
@@ -1,6 +1,10 @@
 //! Typing system to store multiple types in one place with safety
 //! guarantees.
 
+#![no_std]
+
+extern crate alloc;
+
 mod any_index_map;
 mod any_sparse_map;
 pub mod type_pool;
